Add keyboard navigation in table options menu
Refactor drop-down keyboard handling, mainly to be able to reuse the
code. Keyboard handling is moved to a lower level, closer to the DOM,
and operates directly on the rendered elements, instead of on the Vue
component state. This is similar to how Vuetify does it.

This is NOT an example to follow in application code!