Fix bug in SDropDown
Selected items were unselected when removed by text filter. To be able
to fix this, SDropDown needs to implement the internalValue patter, so
now both SDropDown and the internal component does this. It would be
possible to remove this pattern from the internal compoenent, but it's
already there, so I don't see a huge benefit of removing it.

Also, this makes the component look better by hiding the text caret when
the drop-down is closed. I was not able to use the readonly attribute,
because we want it to open if the user starts to type, and not lose the
first character. Using a little trick to hide the caret by setting text
color to transparent.